use new server-discovery everywhere
