+++ packages/core/src/server-discovery/discovery.spec.ts
@@ -1,598 +1,258 @@
-import {
-	afterAll,
-	afterEach,
-	describe,
-	expect,
-	it,
-	jest,
-	mock,
-} from 'bun:test';
+import { describe, expect, it, mock } from 'bun:test';
+import sinon from 'sinon';
 
-/**
- * This is a workaround restore the original module, since bun module mock is a bit weird
- */
-import ResolverModule from 'node:dns/promises';
+const stubs = {
+	fetch: sinon.stub(),
 
-const { Resolver } = ResolverModule;
+	resolveHostname: sinon.stub(),
+	resolveSrv: sinon.stub(),
+} as const;
 
-import {
-	addressWithDefaultPort,
-	getAddressFromTargetWellKnownEndpoint,
-	getWellKnownCachedAddress,
-	isIpLiteral,
-	resolveHostAddressByServerName,
-	resolveUsingSRVRecordsOrFallbackToOtherRecords,
-	resolveWhenServerNameIsAddressWithPort,
-	resolveWhenServerNameIsIpAddress,
-	wellKnownCache,
-} from './discovery';
-
-const mockResolver = {
-	resolveAny: jest.fn(),
-	resolveSrv: jest.fn(),
-};
-
-mock.module('node:dns/promises', () => ({
-	Resolver: jest.fn(() => mockResolver),
+await mock.module('./discovery', () => ({
+	resolveHostname: stubs.resolveHostname,
 }));
 
-afterAll(() => {
-	// restore the original module
-	mock.module('node:dns/promises', () => ({
-		Resolver,
-	}));
-});
+await mock.module('./_resolver', () => ({
+	resolver: {
+		resolveSrv: stubs.resolveSrv,
+	},
+}));
 
-describe('#isIpLiteral()', () => {
-	it('should return true for valid IPv4 addresses', () => {
-		expect(isIpLiteral('192.168.1.1')).toBe(true);
-		expect(isIpLiteral('127.0.0.1')).toBe(true);
+import { _URL } from './_url';
+import { getHomeserverFinalAddress } from './discovery';
+
+const mockFetch = stubs.fetch as unknown as typeof fetch;
+// const originalFetch = globalThis.fetch;
+globalThis.fetch = mockFetch;
+
+// each function describes a stage of the spec to test spec conformity
+// function returns the set of inputs to test with.
+// each step should behave the same way so the modifications to the stub returns should not change.
+//
+
+type INPUT = string;
+type OUTPUT = [`https://${string}:${string | number}`, { Host: string }];
+
+/*
+ * 1. If the hostname is an IP literal, then that IP address should be used, together with the given port number, or 8448 if no port is given. The target server must present a valid certificate for the IP address. The Host header in the request should be set to the server name, including the port if the server name included one.
+ */
+
+function spec_1__1(): [INPUT[], OUTPUT[]] {
+	return [
+		['11.0.0.1', '11.0.0.1:45'],
+		[
+			['https://11.0.0.1:8448' as const, { Host: '11.0.0.1' }],
+			['https://11.0.0.1:45' as const, { Host: '11.0.0.1:45' }],
+		],
+	];
+}
+
+function spec_1__2(): [INPUT[], OUTPUT[]] {
+	return [
+		['[::1]', '[::1]:45'],
+		[
+			['https://[::1]:8448' as const, { Host: '[::1]' }],
+			['https://[::1]:45' as const, { Host: '[::1]:45' }],
+		],
+	];
+}
+
+/*
+ * SPEC:
+ * 2. If the hostname is not an IP literal, and the server name includes an explicit port, resolve the hostname to an IP address using CNAME, AAAA or A records. Requests are made to the resolved IP address and given port with a Host header of the original server name (with port). The target server must present a valid certificate for the hostname.
+ */
+
+function spec_2__1(): [INPUT[], OUTPUT[]] {
+	stubs.resolveHostname.resolves('11.0.0.1');
+	return [
+		['example.com:45'],
+		[['https://11.0.0.1:45' as const, { Host: 'example.com:45' }]],
+	];
+}
+
+function spec_2__2(): [INPUT[], OUTPUT[]] {
+	stubs.resolveHostname.resolves('[::1]');
+	return [
+		['example.com:45'],
+		[['https://[::1]:45' as const, { Host: 'example.com:45' }]],
+	];
+}
+
+// wellknown
+// If <delegated_hostname> is an IP literal, then that IP address should be used together with the <delegated_port> or 8448 if no port is provided. The target server must present a valid TLS certificate for the IP address. Requests must be made with a Host header containing the IP address, including the port if one was provided.
+function spec_3_1__1(): [INPUT[], OUTPUT[]] {
+	// If the hostname is not an IP literal and no port is provided
+	const inputs = ['example.com'];
+
+	stubs.resolveHostname.resolves('11.0.0.1');
+
+	// Mock the .well-known response
+	stubs.fetch.resolves({
+		ok: true,
+		json: () => Promise.resolve({ 'm.server': '11.0.0.1:45' }),
+		headers: new Headers({
+			'cache-control': 'max-age=3600',
+		}),
 	});
 
-	it('should return true for valid IPv6 addresses', () => {
-		expect(isIpLiteral('::1')).toBe(true);
-		expect(isIpLiteral('2001:0db8:85a3:0000:0000:8a2e:0370:7334')).toBe(true);
+	return [inputs, [['https://11.0.0.1:45' as const, { Host: '11.0.0.1:45' }]]];
+}
+
+function spec_3_1__2(): [INPUT[], OUTPUT[]] {
+	const inputs = ['example.com'];
+
+	stubs.resolveHostname.resolves('[::1]');
+
+	stubs.fetch.resolves({
+		ok: true,
+		json: () => Promise.resolve({ 'm.server': '[::1]:45' }),
 	});
 
-	it('should return false for invalid IP addresses', () => {
-		expect(isIpLiteral('999.999.999.999')).toBe(false);
-		expect(isIpLiteral('invalid-ip')).toBe(false);
+	return [inputs, [['https://[::1]:45' as const, { Host: '[::1]:45' }]]];
+}
+
+/* 3.2. If <delegated_hostname> is not an IP literal, and <delegated_port> is present, an IP address is discovered by looking up CNAME, AAAA or A records for <delegated_hostname>. The resulting IP address is used, alongside the <delegated_port>. Requests must be made with a Host header of <delegated_hostname>:<delegated_port>. The target server must present a valid certificate for <delegated_hostname>.
+ */
+function spec_3_2(): [INPUT[], OUTPUT[]] {
+	const inputs = ['example.com'];
+
+	stubs.resolveHostname.reset();
+
+	// for some reason onFirstCall and onSecondCall is not working
+	stubs.resolveHostname.callsFake((hostname: string) => {
+		if (hostname === 'example.com') {
+			return Promise.resolve('11.0.0.1');
+		}
+
+		if (hostname === 'example2.com') {
+			return Promise.resolve('[::1]');
+		}
 	});
 
-	it('should return false for empty string', () => {
-		expect(isIpLiteral('')).toBe(false);
+	stubs.fetch.resolves({
+		ok: true,
+		json: () => Promise.resolve({ 'm.server': 'example2.com:45' }), // delegatedPort is present
 	});
 
-	it('should return false for non-IP strings', () => {
-		expect(isIpLiteral('hello world')).toBe(false);
+	return [inputs, [['https://[::1]:45' as const, { Host: 'example2.com:45' }]]];
+}
+
+/* If <delegated_hostname> is not an IP literal and no <delegated_port> is present, an SRV record is looked up for _matrix-fed._tcp.<delegated_hostname>. This may result in another hostname (to be resolved using AAAA or A records) and port. Requests should be made to the resolved IP address and port with a Host header containing the <delegated_hostname>. The target server must present a valid certificate for <delegated_hostname>.*/
+function spec_3_3__1(): [INPUT[], OUTPUT[]] {
+	const inputs = ['example.com'];
+
+	stubs.resolveHostname.resolves('11.0.0.1');
+
+	stubs.fetch.resolves({
+		ok: true,
+		json: () => Promise.resolve({ 'm.server': 'example2.com' }), // no delegatedPort is present, delegatedHostname is present and not ip
 	});
 
-	it('should return true for IP addresses with ports', () => {
-		expect(isIpLiteral('192.168.1.1:8080')).toBe(true);
-		expect(isIpLiteral('[2001:0db8:85a3:0000:0000:8a2e:0370:7334]:443')).toBe(
-			true,
-		);
-	});
-});
+	stubs.resolveSrv.resolves([{ name: '::1', port: 45 }]);
 
-describe('#addressWithDefaultPort()', () => {
-	it('should append default port 8448 to IPv4 address', () => {
-		expect(addressWithDefaultPort('192.168.1.1')).toBe('192.168.1.1:8448');
-	});
+	return [inputs, [['https://[::1]:45' as const, { Host: 'example2.com' }]]];
+}
 
-	it('should append default port 8448 to IPv6 address', () => {
-		expect(
-			addressWithDefaultPort('[2001:0db8:85a3:0000:0000:8a2e:0370:7334]'),
-		).toBe('[2001:0db8:85a3:0000:0000:8a2e:0370:7334]:8448');
+function spec_3_3__2(): [INPUT[], OUTPUT[]] {
+	const inputs = ['example.com'];
+
+	stubs.resolveHostname.callsFake((name) => {
+		if (name === 'exmaple.com') return '11.0.0.1';
+
+		if (name === 'example3.com') return '[::1]';
 	});
 
-	it('should append default port 8448 to domain name', () => {
-		expect(addressWithDefaultPort('example.com')).toBe('example.com:8448');
-	});
-});
-
-describe('#resolveWhenServerNameIsIpAddress()', () => {
-	it('should return the same address with port if IPv4 address has a port', async () => {
-		const result = await resolveWhenServerNameIsIpAddress('192.168.1.1:8080');
-		expect(result).toBe('192.168.1.1:8080');
+	stubs.fetch.resolves({
+		ok: true,
+		json: () => Promise.resolve({ 'm.server': 'example2.com' }), // no delegatedPort is present, delegatedHostname is present and not ip
 	});
 
-	it('should return the same address with port if IPv6 address has a port', async () => {
-		const result = await resolveWhenServerNameIsIpAddress(
-			'[2001:0db8:85a3:0000:0000:8a2e:0370:7334]:443',
-		);
-		expect(result).toBe('[2001:0db8:85a3:0000:0000:8a2e:0370:7334]:443');
+	stubs.resolveSrv.resolves([{ name: 'example3.com', port: 45 }]); // another hostname
+	// now should do another resolveHostname
+
+	return [inputs, [['https://[::1]:45' as const, { Host: 'example2.com' }]]];
+}
+
+/* If the /.well-known request returned an error response, and no SRV records were found, an IP address is resolved using CNAME, AAAA and A records. Requests are made to the resolved IP address using port 8448 and a Host header containing the <hostname>. The target server must present a valid certificate for <hostname>. */
+function spec_3_4__1(): [INPUT[], OUTPUT[]] {
+	const inputs = ['example.com'];
+
+	stubs.resolveHostname.resolves('11.0.0.1');
+
+	// wellknown no
+	stubs.fetch.resolves({
+		ok: false,
 	});
 
-	it('should append default port 8448 if IPv4 address does not have a port', async () => {
-		const result = await resolveWhenServerNameIsIpAddress('192.168.1.1');
-		expect(result).toBe('192.168.1.1:8448');
+	// srv no
+	stubs.resolveSrv.resolves([]);
+
+	return [
+		inputs,
+		[['https://11.0.0.1:8448' as const, { Host: 'example.com' }]],
+	];
+}
+
+async function runTest(inputs: INPUT[], outputs: OUTPUT[]) {
+	for (let i = 0; i < inputs.length; i++) {
+		const input = inputs[i];
+		const output = outputs[i];
+
+		const [address, headers] = await getHomeserverFinalAddress(input);
+
+		expect(address).toBe(output[0]);
+		expect(headers).toEqual(output[1]);
+	}
+}
+
+describe('_URL', () => {
+	it('should mention port if specified even if standard port is used, unlike node:url', () => {
+		const url = new _URL('https://example.com:443');
+		expect(url.port).toBe('443');
+		expect(url.origin).toBe('https://example.com');
 	});
 
-	it('should append default port 8448 if IPv6 address does not have a port', async () => {
-		const result = await resolveWhenServerNameIsIpAddress(
-			'2001:0db8:85a3:0000:0000:8a2e:0370:7334',
-		);
-		expect(result).toBe('[2001:0db8:85a3:0000:0000:8a2e:0370:7334]:8448');
+	it('should not mention port if not specified, like node:url', () => {
+		const url = new _URL('https://example.com');
+		expect(url.port).toBe('');
+		expect(url.origin).toBe('https://example.com');
 	});
 
-	it('should return the same address if it is a valid IPv4 address with default port', async () => {
-		const result = await resolveWhenServerNameIsIpAddress('192.168.1.1:8448');
-		expect(result).toBe('192.168.1.1:8448');
-	});
-
-	it('should return the same address if it is a valid IPv6 address with default port', async () => {
-		const result = await resolveWhenServerNameIsIpAddress(
-			'[2001:0db8:85a3:0000:0000:8a2e:0370:7334]:8448',
-		);
-		expect(result).toBe('[2001:0db8:85a3:0000:0000:8a2e:0370:7334]:8448');
-	});
-
-	it('should return the provided string with the default port for invalid IP addresses', async () => {
-		const result = await resolveWhenServerNameIsIpAddress('invalid');
-		expect(result).toBe('invalid:8448');
-	});
-
-	it('should throw an error if the URL is not parseable', async () => {
-		await expect(
-			resolveWhenServerNameIsIpAddress('::invalid'),
-		).rejects.toThrow();
-	});
-});
-
-describe('#resolveWhenServerNameIsAddressWithPort()', () => {
-	afterEach(() => {
-		jest.clearAllMocks();
-	});
-
-	it('should resolve to CNAME record with port', async () => {
-		mockResolver.resolveAny.mockResolvedValueOnce([
-			{ type: 'CNAME', value: 'alias.example.com' },
-		]);
-
-		const result =
-			await resolveWhenServerNameIsAddressWithPort('example.com:8080');
-		expect(result).toBe('alias.example.com:8080');
-	});
-
-	it('should resolve to AAAA record with port', async () => {
-		mockResolver.resolveAny.mockResolvedValueOnce([
-			{
-				type: 'AAAA',
-				address: '2001:0db8:85a3:0000:0000:8a2e:0370:7334',
-				ttl: 300,
-			},
-		]);
-
-		const result =
-			await resolveWhenServerNameIsAddressWithPort('example.com:8080');
-		expect(result).toBe('[2001:0db8:85a3:0000:0000:8a2e:0370:7334]:8080');
-	});
-
-	it('should resolve to A record with port', async () => {
-		mockResolver.resolveAny.mockResolvedValueOnce([
-			{ type: 'A', address: '192.168.1.1', ttl: 300 },
-		]);
-
-		const result =
-			await resolveWhenServerNameIsAddressWithPort('example.com:8080');
-		expect(result).toBe('192.168.1.1:8080');
-	});
-
-	it('should return the same address with its port if no records are found', async () => {
-		mockResolver.resolveAny.mockResolvedValueOnce([]);
-
-		const result =
-			await resolveWhenServerNameIsAddressWithPort('example.com:8080');
-		expect(result).toBe('example.com:8080');
-	});
-
-	it('should throw an error if an error occurs during resolution', async () => {
-		mockResolver.resolveAny.mockRejectedValueOnce(
-			new Error('DNS resolution error'),
-		);
-
-		await expect(
-			resolveWhenServerNameIsAddressWithPort('example.com:8080'),
-		).rejects.toThrow('DNS resolution error');
-	});
-});
-
-describe('#resolveUsingSRVRecordsOrFallbackToOtherRecords()', () => {
-	afterEach(() => {
-		jest.clearAllMocks();
-	});
-
-	it('should resolve to SRV record target with default port', async () => {
-		mockResolver.resolveSrv = jest
-			.fn()
-			.mockResolvedValue([
-				{ name: 'srv.example.com', port: 8448, priority: 10, weight: 5 },
-			]);
-		mockResolver.resolveAny.mockResolvedValueOnce([
-			{ type: 'A', address: '192.168.1.1', ttl: 300 },
-		]);
-
-		const result =
-			await resolveUsingSRVRecordsOrFallbackToOtherRecords('example.com');
-		expect(result).toBe('192.168.1.1:8448');
-	});
-
-	it('should resolve to SRV record target with specified port', async () => {
-		mockResolver.resolveSrv = jest
-			.fn()
-			.mockResolvedValue([
-				{ name: 'srv.example.com', port: 8448, priority: 10, weight: 5 },
-			]);
-		mockResolver.resolveAny.mockResolvedValueOnce([
-			{ type: 'A', address: '192.168.1.1', ttl: 300 },
-		]);
-
-		const result =
-			await resolveUsingSRVRecordsOrFallbackToOtherRecords('example.com');
-		expect(result).toBe('192.168.1.1:8448');
-	});
-
-	it('should resolve to AAAA record if no SRV records are found', async () => {
-		mockResolver.resolveSrv = jest.fn().mockResolvedValue([]);
-		mockResolver.resolveAny.mockResolvedValueOnce([
-			{
-				type: 'AAAA',
-				address: '2001:0db8:85a3:0000:0000:8a2e:0370:7334',
-				ttl: 300,
-			},
-		]);
-
-		const result =
-			await resolveUsingSRVRecordsOrFallbackToOtherRecords('example.com');
-		expect(result).toBe('[2001:0db8:85a3:0000:0000:8a2e:0370:7334]:8448');
-	});
-
-	it('should resolve to A record if no SRV records are found', async () => {
-		mockResolver.resolveSrv = jest.fn().mockResolvedValue([]);
-		mockResolver.resolveAny.mockResolvedValueOnce([
-			{ type: 'A', address: '192.168.1.1', ttl: 300 },
-		]);
-
-		const result =
-			await resolveUsingSRVRecordsOrFallbackToOtherRecords('example.com');
-		expect(result).toBe('192.168.1.1:8448');
-	});
-
-	it('should return the same address with default port if no records are found', async () => {
-		mockResolver.resolveSrv = jest.fn().mockResolvedValue([]);
-		mockResolver.resolveAny.mockResolvedValueOnce([]);
-
-		const result =
-			await resolveUsingSRVRecordsOrFallbackToOtherRecords('example.com');
-		expect(result).toBe('example.com:8448');
-	});
-
-	it('should throw an error if an error occurs during SRV resolution', async () => {
-		mockResolver.resolveSrv = jest
-			.fn()
-			.mockRejectedValue(new Error('DNS resolution error'));
-
-		await expect(
-			resolveUsingSRVRecordsOrFallbackToOtherRecords('example.com'),
-		).rejects.toThrow('DNS resolution error');
-	});
-
-	it('should throw an error if an error if an error occurs during ANY resolution', async () => {
-		mockResolver.resolveSrv = jest.fn().mockResolvedValue([]);
-		mockResolver.resolveAny.mockRejectedValueOnce(
-			new Error('DNS resolution error'),
-		);
-
-		await expect(
-			resolveUsingSRVRecordsOrFallbackToOtherRecords('example.com'),
-		).rejects.toThrow('DNS resolution error');
+	it('should parse url without protocol part', () => {
+		const url = new _URL('example.com');
+		expect(url.origin).toBe('https://example.com');
 	});
 });
 
-describe('#getAddressFromTargetWellKnownEndpoint()', () => {
-	afterEach(() => {
-		jest.clearAllMocks();
+describe('[Server Discovery 2.1 - resolve final address] https://spec.matrix.org/v1.12/server-server-api/#resolving-server-names', () => {
+	it('2.1.1 (ipv4)', async () => {
+		return runTest(...spec_1__1());
 	});
-
-	it('should return the address and maxAge from the well-known response', async () => {
-		const mockResponse = {
-			ok: true,
-			json: jest.fn().mockResolvedValueOnce({ 'm.server': 'example.com:8448' }),
-			headers: {
-				get: jest.fn().mockReturnValue('max-age=3600'),
-			},
-		};
-		global.fetch = jest.fn().mockResolvedValueOnce(mockResponse) as any;
-
-		const result = await getAddressFromTargetWellKnownEndpoint('example.com');
-		expect(result).toEqual({ address: 'example.com:8448', maxAge: 3600 });
+	it('2.1.1 (ipv6)', async () => {
+		return runTest(...spec_1__2());
 	});
-
-	it('should return the address with default maxAge if cache-control header is not present', async () => {
-		const mockResponse = {
-			ok: true,
-			json: jest.fn().mockResolvedValue({ 'm.server': 'example.com:8448' }),
-			headers: {
-				get: jest.fn().mockReturnValue(null),
-			},
-		};
-		global.fetch = jest.fn().mockResolvedValueOnce(mockResponse) as any;
-
-		const result = await getAddressFromTargetWellKnownEndpoint('example.com');
-		expect(result).toEqual({ address: 'example.com:8448', maxAge: 86400 });
+	it('2.1.2 (ipv4)', async () => {
+		return runTest(...spec_2__1());
 	});
-
-	it('should throw an error if the well-known response is not ok', async () => {
-		const mockResponse = {
-			ok: false,
-			json: jest.fn(),
-			headers: {
-				get: jest.fn(),
-			},
-		};
-		global.fetch = jest.fn().mockResolvedValueOnce(mockResponse) as any;
-
-		await expect(
-			getAddressFromTargetWellKnownEndpoint('example.com'),
-		).rejects.toThrow('No address found');
+	it('2.1.2 (ipv6)', async () => {
+		return runTest(...spec_2__2());
 	});
-
-	it('should throw an error if json() throws', async () => {
-		const mockResponse = {
-			ok: true,
-			json: jest.fn().mockRejectedValue(new Error('JSON error')),
-			headers: {
-				get: jest.fn(),
-			},
-		};
-		global.fetch = jest.fn().mockResolvedValueOnce(mockResponse) as any;
-
-		await expect(
-			getAddressFromTargetWellKnownEndpoint('example.com'),
-		).rejects.toThrow('No address found');
+	it('3.1.1 (well-known delegation - ip4)', async () => {
+		return runTest(...spec_3_1__1());
 	});
-
-	it('should throw an error if the well-known response does not contain m.server', async () => {
-		const mockResponse = {
-			ok: true,
-			json: jest.fn().mockResolvedValue({}),
-			headers: {
-				get: jest.fn(),
-			},
-		};
-		global.fetch = jest.fn().mockResolvedValueOnce(mockResponse) as any;
-
-		await expect(
-			getAddressFromTargetWellKnownEndpoint('example.com'),
-		).rejects.toThrow('No address found');
+	it('3.1.1 (well-known delegation - ip6)', async () => {
+		return runTest(...spec_3_1__2());
 	});
-
-	it('should limit maxAge to 48 hours if cache-control header exceeds the limit', async () => {
-		const mockResponse = {
-			ok: true,
-			json: jest.fn().mockResolvedValue({ 'm.server': 'example.com:8448' }),
-			headers: {
-				get: jest.fn().mockReturnValue('max-age=200000'),
-			},
-		};
-		global.fetch = jest.fn().mockResolvedValueOnce(mockResponse) as any;
-
-		const result = await getAddressFromTargetWellKnownEndpoint('example.com');
-		expect(result).toEqual({ address: 'example.com:8448', maxAge: 172800 });
+	it('3.2.1 (well-known delegation)', async () => {
+		return runTest(...spec_3_2());
 	});
-});
-
-describe('#getWellKnownCachedAddress()', () => {
-	afterEach(() => {
-		jest.clearAllMocks();
+	it('3.3.1 (well-known delegation)', async () => {
+		return runTest(...spec_3_3__1());
 	});
-
-	it('should return the cached address if it is still valid', () => {
-		const serverName = 'example.com';
-		const cachedData = {
-			address: 'cached.example.com:8448',
-			maxAge: 3600,
-			timestamp: Date.now() - 1000, // 1 second ago
-		};
-		wellKnownCache.set(serverName, cachedData);
-
-		const result = getWellKnownCachedAddress(serverName);
-		expect(result).toBe('cached.example.com:8448');
+	it('3.3.2 (well-known delegation)', async () => {
+		return runTest(...spec_3_3__2());
 	});
-
-	it('should return null if the cached address has expired', () => {
-		const serverName = 'example.com';
-		const cachedData = {
-			address: 'cached.example.com:8448',
-			maxAge: 3600,
-			timestamp: Date.now() - 4000 * 1000, // 4000 seconds ago
-		};
-		wellKnownCache.set(serverName, cachedData);
-
-		const result = getWellKnownCachedAddress(serverName);
-		expect(result).toBeNull();
-	});
-
-	it('should return null if there is no cached address for the server name', () => {
-		const serverName = 'nonexistent.com';
-
-		const result = getWellKnownCachedAddress(serverName);
-		expect(result).toBeNull();
-	});
-});
-
-describe('#resolveHostAddressByServerName()', () => {
-	const localHomeServerName = 'rc1';
-	const localHomeServerNameWithPort = 'rc1:443';
-
-	afterEach(() => {
-		wellKnownCache.clear();
-	});
-
-	it('should resolve IP literal addresses directly', async () => {
-		const { address, headers } = await resolveHostAddressByServerName(
-			'192.168.1.1',
-			localHomeServerName,
-		);
-		expect(address).toBe('192.168.1.1:8448');
-		expect(headers).toEqual({ Host: localHomeServerNameWithPort });
-	});
-
-	it('should resolve addresses with explicit ports directly', async () => {
-		mockResolver.resolveAny.mockResolvedValueOnce([
-			{
-				type: 'AAAA',
-				address: '2001:0db8:85a3:0000:0000:8a2e:0370:7334',
-				ttl: 300,
-			},
-		]);
-		const { address, headers } = await resolveHostAddressByServerName(
-			'example.com:8080',
-			localHomeServerName,
-		);
-		expect(address).toBe('[2001:0db8:85a3:0000:0000:8a2e:0370:7334]:8080');
-		expect(headers).toEqual({ Host: localHomeServerNameWithPort });
-	});
-
-	// TODO: Make it pass again
-	it.skip('should return cached address if available and valid', async () => {
-		mockResolver.resolveAny.mockResolvedValueOnce([
-			{
-				type: 'AAAA',
-				address: '2001:0db8:85a3:0000:0000:8a2e:0370:7334',
-				ttl: 300,
-			},
-		]);
-		const serverName = 'example.com';
-		const cachedData = {
-			address: 'cached.example.com:8448',
-			maxAge: 3600,
-			timestamp: Date.now() - 1000, // 1 second ago
-		};
-		wellKnownCache.set(serverName, cachedData);
-
-		const { address, headers } = await resolveHostAddressByServerName(
-			serverName,
-			localHomeServerName,
-		);
-		expect(address).toBe('[2001:0db8:85a3:0000:0000:8a2e:0370:7334]:8448');
-		expect(headers).toEqual({ Host: 'cached.example.com:8448' });
-	});
-
-	// TODO: Make it pass again
-	it.skip('should resolve using well-known address if not cached', async () => {
-		mockResolver.resolveAny.mockResolvedValueOnce([
-			{
-				type: 'AAAA',
-				address: '2001:0db8:85a3:0000:0000:8a2e:0370:7334',
-				ttl: 300,
-			},
-		]);
-		const mockResponse = {
-			ok: true,
-			json: jest.fn().mockResolvedValue({ 'm.server': 'example.com:8448' }),
-			headers: {
-				get: jest.fn().mockReturnValue('max-age=3600'),
-			},
-		};
-		global.fetch = jest.fn().mockResolvedValueOnce(mockResponse) as any;
-
-		const { address, headers } = await resolveHostAddressByServerName(
-			'example.com',
-			localHomeServerName,
-		);
-		expect(address).toBe('[2001:0db8:85a3:0000:0000:8a2e:0370:7334]:8448');
-		expect(headers).toEqual({ Host: 'example.com:8448' });
-	});
-
-	it('should fallback to SRV records if well-known address is not available', async () => {
-		global.fetch = jest
-			.fn()
-			.mockRejectedValueOnce(new Error('Fetch error')) as any;
-		mockResolver.resolveSrv = jest.fn().mockResolvedValue([]);
-		mockResolver.resolveAny.mockResolvedValueOnce([
-			{ type: 'A', address: '192.168.1.1', ttl: 300 },
-		]);
-
-		const { address, headers } = await resolveHostAddressByServerName(
-			'example.com',
-			localHomeServerName,
-		);
-		expect(address).toBe('192.168.1.1:8448');
-		expect(headers).toEqual({ Host: 'example.com' });
-	});
-
-	it('should return the provided address with the default port when the request did not throw but ok is false', async () => {
-		mockResolver.resolveSrv = jest.fn().mockResolvedValue([]);
-		mockResolver.resolveAny.mockResolvedValueOnce([]);
-
-		const mockResponse = {
-			ok: false,
-			json: jest.fn(),
-			headers: {
-				get: jest.fn(),
-			},
-		};
-		global.fetch = jest.fn().mockResolvedValueOnce(mockResponse) as any;
-
-		const { address, headers } = await resolveHostAddressByServerName(
-			'example.com',
-			localHomeServerName,
-		);
-		expect(address).toBe('example.com:8448');
-		expect(headers).toEqual({ Host: 'example.com' });
-	});
-
-	it('should return the provided address with the default port when the request did not throw but json() threw', async () => {
-		mockResolver.resolveSrv = jest.fn().mockResolvedValue([]);
-		mockResolver.resolveAny.mockResolvedValueOnce([]);
-
-		const mockResponse = {
-			ok: true,
-			json: jest.fn().mockRejectedValue(new Error('JSON error')),
-			headers: {
-				get: jest.fn(),
-			},
-		};
-		global.fetch = jest.fn().mockResolvedValueOnce(mockResponse) as any;
-
-		const { address, headers } = await resolveHostAddressByServerName(
-			'example.com',
-			localHomeServerName,
-		);
-		expect(address).toBe('example.com:8448');
-		expect(headers).toEqual({ Host: 'example.com' });
-	});
-
-	it('should fallback to default port if no SRV, CNAME, AAAA, not A records are found', async () => {
-		global.fetch = jest
-			.fn()
-			.mockRejectedValueOnce(new Error('Fetch error')) as any;
-		mockResolver.resolveSrv = jest.fn().mockResolvedValue([]);
-		mockResolver.resolveAny.mockResolvedValueOnce([]);
-
-		const { address, headers } = await resolveHostAddressByServerName(
-			'example.com',
-			localHomeServerName,
-		);
-		expect(address).toBe('example.com:8448');
-		expect(headers).toEqual({ Host: 'example.com' });
-	});
-
-	it('should handle errors gracefully and return address with default port', async () => {
-		global.fetch = jest
-			.fn()
-			.mockRejectedValueOnce(new Error('Fetch error')) as any;
-		mockResolver.resolveSrv = jest
-			.fn()
-			.mockRejectedValue(new Error('DNS resolution error'));
-
-		const { address, headers } = await resolveHostAddressByServerName(
-			'example.com',
-			localHomeServerName,
-		);
-		expect(address).toBe('example.com:8448');
-		expect(headers).toEqual({ Host: 'example.com' });
+	it('3.4.1 (well-known delegation - no wellknown, no srv)', async () => {
+		return runTest(...spec_3_4__1());
 	});
 });
