+++ packages/federation-sdk/tsconfig.json
@@ -8,6 +8,13 @@
 		"verbatimModuleSyntax": false,
 		"tsBuildInfoFile": "./.tsbuildinfo"
 	},
-	"include": ["src/**/*"],
-	"exclude": ["node_modules", "dist", "**/*.spec.ts", "**/*.test.ts"]
+	"include": [
+		"src/**/*",
+	],
+	"exclude": [
+		"node_modules",
+		"dist",
+		"**/*.spec.ts",
+		"**/*.test.ts"
+	]
 }
